+++ packages/transaction-controller/src/types.ts
@@ -268,6 +268,9 @@ export type TransactionMeta = {
   /** Whether MetaMask will be compensated for the gas fee by the transaction. */
   isGasFeeIncluded?: boolean;
 
+  /** Whether the intent of the transaction was achieved via an alternate route or chain. */
+  isIntentComplete?: boolean;
+
   /**
    * Whether the transaction is an incoming token transfer.
    */
@@ -1434,7 +1437,7 @@ export type Layer1GasFeeFlow = {
    * @param args - The arguments for the matcher function.
    * @param args.transactionMeta - The transaction metadata.
    * @param args.messenger - The messenger instance.
-   * @returns Whether the gas fee flow supports the transaction.
+   * @returns A promise that resolves to whether the gas fee flow supports the transaction.
    */
   matchesTransaction({
     transactionMeta,
@@ -1442,7 +1445,7 @@ export type Layer1GasFeeFlow = {
   }: {
     transactionMeta: TransactionMeta;
     messenger: TransactionControllerMessenger;
-  }): boolean;
+  }): Promise<boolean>;
 
   /**
    * Get layer 1 gas fee estimates for a specific transaction.